refactor: :recycle: Split up cli code from api.py (#358)
* refactor: remove api code from cli.py

* refactor(api): :recycle: break out cli code from api.py

* cli arguments now should go directly to mintapi instead of api.py. See README.md

* doc: :memo: Edit README.md to update for changes

refs: #358

* doc: :memo: fix setup.py long description to point to readme

* refactor: :recycle: synchronize loggers
