+++ mintapi/__init__.py
@@ -1,6 +1,6 @@
-from .api import Mint
-from .api import get_accounts
-from .api import get_budgets
-from .api import initiate_account_refresh
-from .api import main
-from .api import print_accounts
+import logging
+
+from mintapi.api import *
+
+
+logging.getLogger("mintapi").setLevel(logging.INFO)
