+++ mintapi/__main__.py
@@ -1,0 +1,5 @@
+from mintapi.cli import main
+
+
+if __name__ == "__main__":
+    main()
